+++ src/h5radiomics/utils/feature_utils.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Dict, Any
