+++ src/h5radiomics/utils/maskgeo.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import math
 import os 
 from typing import Tuple, Optional
